% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractify.R
\name{abstractify}
\alias{abstractify}
\title{Abstract highlighted text to new function call}
\usage{
abstractify()
}
\description{
Sometimes you write a bunch of lines of code and decide that it should
actually be a function so that you can reuse it with different arguments.
You would need to identify which symbols in the code need to be abstracted
over and added to the list of arguments. It can be tedious to hunt down
everything, so this function will do it for you by identifying all the
symbols in some highlighted text and finding which ones are not accounted
for by function calls or builtin operators in the global environment.
}
\details{
This function will add text to the currently active source document,
reformat it, then highlight and shift focus to the new function name.
}
