% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.get_abstraction_candidates}
\alias{.get_abstraction_candidates}
\title{Check type of identified symbols}
\usage{
.get_abstraction_candidates(global_symbols)
}
\arguments{
\item{global_symbols}{Output of \code{.get_global_symbols}}
}
\value{
Character vector of length <= global_symbols (usually smaller)
}
\description{
Figure out which symbols we need to extract over.
Some things not to abstract over:
\itemize{
\item closure = function call
\item special = things like ::
\itemize{
\item error   = typically from tidyverse verbs where a symbol is to be
evaluated in the context of a piped in dataframe, eg cyl in
mutate(mtcars, cyl = 3)
}
\item builtin  = things like c or \code{+}
}
}
