% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.get_assigned_symnames}
\alias{.get_assigned_symnames}
\title{Get LHS of assignment}
\usage{
.get_assigned_symnames(selected_exprs)
}
\arguments{
\item{selected_exprs}{Output of \code{parse(text = text)}}
}
\value{
Character vector of length 1 containing the LHS of an assignment
}
\description{
Get LHS of assignment
}
