% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.get_global_symbols}
\alias{.get_global_symbols}
\title{Find which symbols exist in global environment}
\usage{
.get_global_symbols(selected_exprs)
}
\arguments{
\item{selected_exprs}{Output of \code{parse(text = text)}}
}
\value{
Character vector of symbol names
}
\description{
Used to find the unique set of symbols from a provided set of expressions
in the global environment
}
