% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracted_text.R
\name{.get_highlighted_text}
\alias{.get_highlighted_text}
\title{Get highlighted text only}
\usage{
.get_highlighted_text()
}
\value{
Highlighted text as a string
}
\description{
Helper for testing
}
