% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.get_ignored_syms}
\alias{.get_ignored_syms}
\title{Determine which symbols to ignore}
\usage{
.get_ignored_syms(selected_exprs)
}
\arguments{
\item{selected_exprs}{Output of \code{parse(text = text)}}
}
\value{
Logical named vector of which symbols should be ignored. Note that
this is later used by indexing via the symbol name, and if it doesn't exist
then NA is returned which is picked up by a call to \code{is.na}
}
\description{
Symbols should be ignored if they're on the LHS of an assigment
}
