% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.is_assignment_expr}
\alias{.is_assignment_expr}
\title{Check if expression is an assignment}
\usage{
.is_assignment_expr(selected_exprs)
}
\arguments{
\item{selected_exprs}{Output of \code{parse(text = text)}}
}
\value{
Logical vector of which expressions are assignments
}
\description{
Checks if any expressions are of the form \code{varname <- ...}, in which
case we don't need to care about varname.
}
