% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.parse_symbols}
\alias{.parse_symbols}
\title{Parse the highlighted text and get all the global variables}
\usage{
.parse_symbols(text)
}
\arguments{
\item{text}{Text from \code{getSourceEditorContext()$selection}}
}
\value{
A character vector of which symbols should be abstracted over for
a new function definition
}
\description{
Wrapper for the symbol extraction helpers
}
