% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_handling_helpers.R
\name{.remove_assigned_candidates}
\alias{.remove_assigned_candidates}
\title{Filter out assignments}
\usage{
.remove_assigned_candidates(selected_exprs, abstraction_candidates)
}
\arguments{
\item{selected_exprs}{Output of \code{parse(text = text)}}

\item{abstraction_candidates}{Output of \code{.get_abstraction_candidates}}
}
\value{
Character vector of length <= abstraction candidates (smaller if
there were any assignment operations)
}
\description{
Removes any candidate symbols on the LHS of an assignment expression
}
