% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracted_text.R
\name{get_abstracted_text}
\alias{get_abstracted_text}
\title{Abstract a function based on global environment}
\usage{
get_abstracted_text(text)
}
\arguments{
\item{text}{Text from \code{getSourceEditorContext()$selection}}
}
\value{
A function definition of the following form:

new_function <- function(..identified symbols..) {
...highlighted text...
}
}
\description{
Given highlighted text from \code{getSourceEditorContext()$selection}, determine
which symbols are in the global environment and should be abstracted over
as a function call argument. This will ignore any symbols corresponding to
function calls, operators, unquoted vars in tidyverse verbs, and builtin
symbols.
}
