% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ending_position.R
\name{get_ending_position}
\alias{get_ending_position}
\title{Get last non-empty row from selection}
\usage{
get_ending_position(text, end_row, start_row)
}
\arguments{
\item{text}{Text from \code{getSourceEditorContext()$selection}}

\item{end_row}{Last row index in selection}

\item{start_row}{First row index in selection, for offset}
}
\value{
Integer of 1 + Index of last non-empty row
}
\description{
Ignores spaces and tabs
}
