% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_abstracted_text.R
\name{insert_abstracted_text}
\alias{insert_abstracted_text}
\title{Insert and reformat function definition}
\usage{
insert_abstracted_text(fx_text, start_row, new_end_row, doc_id)
}
\arguments{
\item{fx_text}{Text to insert, a function definition}

\item{start_row}{Starting row of selection}

\item{new_end_row}{Ending row for insertion, from \code{get_ending_position}}

\item{doc_id}{Active document id returned from \code{getSourceEditorContext()}}
}
\description{
Insert and reformat function definition
}
